build: remove commit hook for git-cz
"Prepare commit message" hook runs not only on commit, but also on a variety of other git tasks such
as rebase and merge, sometimes in an automated way from IDE (Jetbrains IDEs do that). git-cz utility
wasn't built with this in mind, and leads to bad developer experience. We can use already existing
`npm run commit` script for this. Also `"prepare"` npm hook runs when a package is packaged (`npm
pack`), and `husky` is intended to be run on install.